Fix UTM band calculation to use int conversion (#1)
convert_wgs_to_utm in humminbird_class.py and lowrance_class.py

If you input the coordinates from your file it points to Antarctica because np.floor() converts the output to a float so len(utm_band) can't = 1. epsg_code = 3263.0 which is in Antarctica. Output should be epsg_code = 32603. The "03" is just "3". UTM zones 1-9 are not working correctly.https://github.com/maxschwartziv/PINGVerter_UTM1to9Fix/blob/main/README.md

https://www.online-python.com/z23QrCD4Kc

If we change the function from
utm_band = str((np.floor((lon + 180) / 6 ) % 60) + 1)

to

utm_band = str(int(np.floor((lon + 180) / 6 ) % 60) + 1)

convert_wgs_to_utm should work as intended.

+++ pingverter/humminbird_class.py
@@ -414,7 +414,7 @@ class hum(object):
         This function estimates UTM zone from geographic coordinates
         see https://stackoverflow.com/questions/40132542/get-a-cartesian-projection-accurate-around-a-lat-lng-pair
         """
-        utm_band = str((np.floor((lon + 180) / 6 ) % 60) + 1)
+        utm_band = str(int(np.floor((lon + 180) / 6 ) % 60) + 1)
         if len(utm_band) == 1:
             utm_band = '0'+utm_band
         if lat >= 0:
